Use enet connectID instead of rolling our own ID

+++ src/slap42/networking/message_types.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <string>
 #include <bytepack/bytepack.hpp>
 #include <glm/glm.hpp>
@@ -14,7 +15,7 @@ enum class MessageType : uint8_t {
   kChatMessage         = 0x04,
 };
 
-typedef uint8_t peer_id;
+typedef uint32_t peer_id;
 
 struct PlayerPositionUpdateMessage {
   peer_id id;
